+++ rust/dbn/src/record/conv.rs
@@ -124,9 +124,15 @@ pub fn str_to_c_chars<const N: usize>(s: &str) -> Result<[c_char; N]> {
 /// None.
 ///
 /// # Errors
-/// This function returns an error if `chars` contains invalid UTF-8.
+/// This function returns an error if `chars` contains invalid UTF-8 or is not null-terminated.
 pub fn c_chars_to_str<const N: usize>(chars: &[c_char; N]) -> Result<&str> {
-    let cstr = unsafe { CStr::from_ptr(chars.as_ptr()) };
+    // Safety: Casting from i8 to u8 slice should be safe
+    let bytes = unsafe { as_u8_slice(chars) };
+    let cstr = CStr::from_bytes_until_nul(bytes).map_err(|_| Error::Conversion {
+        input: format!("{chars:?}"),
+        desired_type: "CStr (null-terminated)",
+    })?;
+
     cstr.to_str()
         .map_err(|e| Error::utf8(e, format!("converting c_char array: {chars:?}")))
 }
@@ -166,5 +172,39 @@ pub(crate) mod cstr_serde {
     {
         let str = String::deserialize(deserializer)?;
         str_to_c_chars(&str).map_err(de::Error::custom)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::os::raw::c_char;
+
+    #[test]
+    fn test_c_chars_to_str_success() {
+        let null_terminated: [c_char; 5] = [
+            'A' as c_char,
+            'A' as c_char,
+            'A' as c_char,
+            'A' as c_char,
+            0,
+        ];
+        let result = c_chars_to_str(&null_terminated);
+        assert_eq!(result.unwrap(), "AAAA");
+    }
+
+    #[test]
+    fn test_c_chars_to_str_failure_on_missing_null_terminator() {
+        let non_null_terminated: [c_char; 5] = ['A' as c_char; 5];
+        let err = c_chars_to_str(&non_null_terminated)
+            .expect_err("Expected failure on non-null terminated string");
+
+        assert!(matches!(
+            err,
+            Error::Conversion {
+                input: _,
+                desired_type: "CStr (null-terminated)",
+            }
+        ));
     }
 }